backend/pyproject.toml
[tool.poetry]
name = "cras-backend"
version = "0.1.0"
description = "Clinical Research Management System (CRAS) Backend"
authors = ["HKU Research <research@hku.hk>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
sqlmodel = "0.0.12"
alembic = "^1.12.1"
uvicorn = "^0.24.0"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
httpx = "^0.25.1"
uuid6 = "^2024.1.12"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
